+++ scripts/translate/mod.ts
@@ -47,8 +47,8 @@ async function completeTranslationsFor(localesDir: string) {
 await completeTranslationsFor("src/ui/react/toolbar/i18n/translations");
 await completeTranslationsFor("src/ui/react/weg/i18n/translations");
 await completeTranslationsFor("src/ui/react/settings/i18n/translations");
-await completeTranslationsFor("src/ui/react/wallpaper_manager/i18n/translations");
 
+await completeTranslationsFor("src/ui/svelte/wallpaper_manager/i18n/translations");
 await completeTranslationsFor("src/ui/svelte/power-menu/i18n/translations");
 await completeTranslationsFor("src/ui/svelte/bluetooth-popup/i18n/translations");
 await completeTranslationsFor("src/ui/svelte/network-popup/i18n/translations");